+++ presentation/src/main/java/com/dcs/presentation/ui/people/PeopleEffect.kt
@@ -1,5 +1,5 @@
 package com.dcs.presentation.ui.people
 
 sealed interface PeopleEffect {
-    data class NavigateToDetail(val personId: Long) : PeopleEffect
+    data class NavigateToDetail(val personId: Int) : PeopleEffect
 }
